X & Y as ratio instead of px
* Changed field types of X & Y from Integer values to Float values (from 0 to 1 in mind, ratio)
* Added event listener for window resize, it means that content is being resized every time page changes its size
* Reduced the amount of triggering setState for person coordinates (once per getCoordinates function instead of once per container)
* Changed calcs to use ratio value together w/ window size, instead of readed coord from API
* Save button now saves the values in ratio form, instead of px (coords)
* Deleted unused 'reset' variable from setState
* Cleaned up some code

+++ django-react/backend/familytree/models.py
@@ -32,8 +32,8 @@ class FamilytreePerson(models.Model):
     sex_choices = [(male, 'male'),(female, 'female'),(other, 'other')]
     sex_choices = models.CharField(max_length = 35,choices = sex_choices, default = other)
 
-    x = models.IntegerField(default=0)
-    y = models.IntegerField(default=0)
+    x = models.FloatField(default=0)
+    y = models.FloatField(default=0)
 
     def _str_(self):
         return self.first_name + " " + self.last_name
